requirements.txt
prettytable
python-ldap